+++ apps/ekstra-omsorgsdager-andre-forelder-ikke-tilsyn/src/app/søknad/SøknadStep.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { useLogSidevisning } from '@navikt/sif-common-amplitude';
 import { soknadStepUtils, Step } from '@navikt/sif-common-soknad-ds';
 import InvalidStepSøknadsdataInfo from '../components/invalid-step-søknadsdata-info/InvalidStepSøknadsdataInfo';
 import useAvbrytEllerFortsettSenere from '../hooks/useAvbrytSøknad';
@@ -14,8 +13,6 @@ interface Props {
 
 const SøknadStep: React.FunctionComponent<Props> = ({ stepId, children }) => {
     const { intl, text } = useAppIntl();
-
-    useLogSidevisning(stepId);
 
     const stepConfig = getSøknadStepConfig();