+++ apps/omsorgspengerutbetaling-soknad/src/app/pages/kvittering/KvitteringPage.tsx
@@ -1,5 +1,4 @@
 import { Heading, Link } from '@navikt/ds-react';
-import { SIFCommonPageKey, useLogSidevisning } from '@navikt/sif-common-amplitude';
 import Block from '@navikt/sif-common-core-ds/src/atoms/block/Block';
 import CheckmarkIcon from '@navikt/sif-common-core-ds/src/atoms/checkmark-icon/CheckmarkIcon';
 import Checklist from '@navikt/sif-common-core-ds/src/components/lists/checklist/Checklist';
@@ -20,8 +19,6 @@ const KvitteringPage = ({ onUnmount }: Props) => {
             onUnmount();
         };
     });
-
-    useLogSidevisning(SIFCommonPageKey.kvittering);
 
     return (
         <Page title={text('page.kvittering.sidetittel')}>