+++ apps/pleiepenger-i-livets-sluttfase-soknad/src/app/søknad/SøknadStep.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { useLogSidevisning } from '@navikt/sif-common-amplitude';
 import { soknadStepUtils, Step } from '@navikt/sif-common-soknad-ds';
 import InvalidStepSøknadsdataInfo from '../components/invalid-step-søknadsdata-info/InvalidStepSøknadsdataInfo';
 import useAvbrytEllerFortsettSenere from '../hooks/useAvbrytSøknad';
@@ -18,8 +17,6 @@ const SøknadStep: React.FunctionComponent<Props> = ({ stepId, children }) => {
     const {
         state: { søknadsdata },
     } = useSøknadContext();
-
-    useLogSidevisning(stepId);
 
     const stepConfig = getSøknadStepConfig(søknadsdata);