+++ apps/sif-ettersending/src/app/pages/ikke-tilgang-page/ikkeTilgangPage.tsx
@@ -1,4 +1,3 @@
-import { SIFCommonPageKey, useLogSidevisning } from '@navikt/sif-common-amplitude';
 import Block from '@navikt/sif-common-core-ds/src/atoms/block/Block';
 import Page from '@navikt/sif-common-core-ds/src/components/page/Page';
 import SifGuidePanel from '@navikt/sif-common-core-ds/src/components/sif-guide-panel/SifGuidePanel';
@@ -12,7 +11,6 @@ interface Props {
 
 const IkkeTilgangPage = ({ søknadstype }: Props) => {
     const { text } = useAppIntl();
-    useLogSidevisning(SIFCommonPageKey.ikkeTilgang);
 
     return (
         <Page